backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.39
alembic==1.15.1
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.11.1
pydantic-settings==2.8.1
python-multipart==0.0.20
email-validator==2.2.0
